+++ data/src/main/java/telegram4j/json/MessageIdData.java
@@ -15,5 +15,5 @@ public interface MessageIdData {
     }
 
     @JsonProperty("message_id")
-    int messageId();
+    long messageId();
 }
